refactor(rebuild): move N1x authority decision
